+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 **Please note that Webpacker 3.1.0 and 3.1.1 has some serious bugs so please consider using either 3.0.2 or 3.2.0**
+## Unreleased
+
+### Breaking changes
+
+- Fixes #1281 by installing binstubs only as local executables. To upgrade, run the install task and make sure the executables in the `bin` directory are overwritten.
 
 ## [3.2.2] - 2018-02-11